moved the utils methods to helpers folder
